backend/requirements.txt
fastapi
uvicorn
numpy
Pillow
PyPDF2
python-multipart